64bit compile fixes

+++ hw/nand.c
@@ -589,7 +589,7 @@ static void glue(nand_blk_write_, PAGE_SIZE)(NANDFlashState *s)
         off = (s->addr & PAGE_MASK) + s->offset;
         soff = SECTOR_OFFSET(s->addr);
         if (bdrv_read(s->bdrv, sector, iobuf, PAGE_SECTORS) == -1) {
-            printf("%s: read error in sector %lli\n", __FUNCTION__, sector);
+            printf("%s: read error in sector %llu\n", __FUNCTION__, (long long unsigned)sector);
             return;
         }
 
@@ -609,13 +609,13 @@ static void glue(nand_blk_write_, PAGE_SIZE)(NANDFlashState *s)
         }
 
         if (bdrv_write(s->bdrv, sector, iobuf, PAGE_SECTORS) == -1)
-            printf("%s: write error in sector %lli\n", __FUNCTION__, sector);
+            printf("%s: write error in sector %llu\n", __FUNCTION__, (long long unsigned)sector);
     } else {
         off = PAGE_START(s->addr) + (s->addr & PAGE_MASK) + s->offset;
         sector = off >> 9;
         soff = off & 0x1ff;
         if (bdrv_read(s->bdrv, sector, iobuf, PAGE_SECTORS + 2) == -1) {
-            printf("%s: read error in sector %lli\n", __FUNCTION__, sector);
+            printf("%s: read error in sector %llu\n", __FUNCTION__, (long long unsigned)sector);
             return;
         }
 
@@ -626,7 +626,7 @@ static void glue(nand_blk_write_, PAGE_SIZE)(NANDFlashState *s)
         }
 
         if (bdrv_write(s->bdrv, sector, iobuf, PAGE_SECTORS + 2) == -1)
-            printf("%s: write error in sector %lli\n", __FUNCTION__, sector);
+            printf("%s: write error in sector %llu\n", __FUNCTION__, (long long unsigned)sector);
     }
     s->offset = 0;
 }
@@ -651,29 +651,29 @@ static void glue(nand_blk_erase_, PAGE_SIZE)(NANDFlashState *s)
         page = SECTOR(addr + (ADDR_SHIFT + s->erase_shift));
         for (; i < page; i ++)
             if (bdrv_write(s->bdrv, i, iobuf, 1) == -1)
-                printf("%s: write error in sector %lli\n", __FUNCTION__, i);
+                printf("%s: write error in sector %llu\n", __FUNCTION__, (long long unsigned)i);
     } else {
         addr = PAGE_START(addr);
         page = addr >> 9;
         if (bdrv_read(s->bdrv, page, iobuf, 1) == -1)
-            printf("%s: read error in sector %lli\n", __FUNCTION__, page);
+            printf("%s: read error in sector %llu\n", __FUNCTION__, (long long unsigned)page);
         memset(iobuf + (addr & 0x1ff), 0xff, (~addr & 0x1ff) + 1);
         if (bdrv_write(s->bdrv, page, iobuf, 1) == -1)
-            printf("%s: write error in sector %lli\n", __FUNCTION__, page);
+            printf("%s: write error in sector %llu\n", __FUNCTION__, (long long unsigned)page);
 
         memset(iobuf, 0xff, 0x200);
         i = (addr & ~0x1ff) + 0x200;
         for (addr += ((PAGE_SIZE + OOB_SIZE) << s->erase_shift) - 0x200;
                         i < addr; i += 0x200)
             if (bdrv_write(s->bdrv, i >> 9, iobuf, 1) == -1)
-                printf("%s: write error in sector %lli\n", __FUNCTION__, i >> 9);
+                printf("%s: write error in sector %llu\n", __FUNCTION__, (long long unsigned)i >> 9);
 
         page = i >> 9;
         if (bdrv_read(s->bdrv, page, iobuf, 1) == -1)
-            printf("%s: read error in sector %lli\n", __FUNCTION__, page);
+            printf("%s: read error in sector %llu\n", __FUNCTION__, (long long unsigned)page);
         memset(iobuf, 0xff, ((addr - 1) & 0x1ff) + 1);
         if (bdrv_write(s->bdrv, page, iobuf, 1) == -1)
-            printf("%s: write error in sector %lli\n", __FUNCTION__, page);
+            printf("%s: write error in sector %llu\n", __FUNCTION__, (long long unsigned)page);
     }
 }
 
@@ -686,8 +686,8 @@ static void glue(nand_blk_load_, PAGE_SIZE)(NANDFlashState *s,
     if (s->bdrv) {
         if (s->mem_oob) {
             if (bdrv_read(s->bdrv, SECTOR(addr), s->io, PAGE_SECTORS) == -1)
-                printf("%s: read error in sector %lli\n",
-                                __FUNCTION__, SECTOR(addr));
+                printf("%s: read error in sector %llu\n",
+                                __FUNCTION__, (long long unsigned)SECTOR(addr));
             memcpy(s->io + SECTOR_OFFSET(s->addr) + PAGE_SIZE,
                             s->storage + (PAGE(s->addr) << OOB_SHIFT),
                             OOB_SIZE);
@@ -695,8 +695,8 @@ static void glue(nand_blk_load_, PAGE_SIZE)(NANDFlashState *s,
         } else {
             if (bdrv_read(s->bdrv, PAGE_START(addr) >> 9,
                                     s->io, (PAGE_SECTORS + 2)) == -1)
-                printf("%s: read error in sector %lli\n",
-                                __FUNCTION__, PAGE_START(addr) >> 9);
+                printf("%s: read error in sector %llu\n",
+                                __FUNCTION__, (long long unsigned)PAGE_START(addr) >> 9);
             s->ioaddr = s->io + (PAGE_START(addr) & 0x1ff) + offset;
         }
     } else {
